+++ tests/test_vsim.py
@@ -37,7 +37,7 @@ simulation = VSim(
 #   driver settings
 # ----------------------------------------------------------------------------
 
-from mesa.driver import GeneticOptimizer
+from mesa.drivers import GeneticOptimizer
 
 # parameters to fix or vary. To fix set as a single value. To vary set bounds as a tuple
 params = {
